requirements.txt
streamlit
streamlit-lottie
requests
streamlit-custom-toggle
pillow